% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.outlier.R
\name{check.outlier}
\alias{check.outlier}
\title{Checking outliers}
\usage{
check.outlier(y, X = NULL, U = NULL, contrast = list(mean = NULL,
  dispersion = NULL), offsets = NULL, check = "contrast",
  cut.off = outlier.cutoff(0.05), ZI = TRUE, conserv = TRUE,
  verbose = TRUE, output = FALSE, ...)
}
\arguments{
\item{y}{vector containing the counts of a gene.}

\item{X}{vector or matrix containing covariates in the mean GLM.}

\item{U}{vector or matrix containing covariates in the dispersion GLM, 
which can be different from X.}

\item{contrast}{list containing contrast matices for mean and dispersion.}

\item{offsets}{numeric vector of same size as y giving offsets for use in log-linear models.}

\item{check}{which covariate(covariates) to check. if check="contrast", outliers 
influential on the effects of contrasts will be checked; if check="other", outliers 
influential on the effects of all other covariates (except contrasts) will be checked; 
if check="all", outliers influential on the effects of all covariates will be checked. 
In addition, outliers influential towards specific covariate(/covariates) can be checked. 
User needs to provide the name(s) of targeted covariate(s).}

\item{cut.off}{cutoff point defining the significance level of outlier selections.}

\item{ZI}{logical, if TRUE (default), the effect of zero-inflation will be considered.}

\item{conserv}{logical, if TRUE (default), outlier detection will skip any gene with estimation 
error for either coefficients or standard errors of coefficients and label it as an error; 
if FALSE, outlier detection will only skip any gene with estimation error 
on coefficients.}

\item{verbose}{logical, if TRUE, progress information will be reported.}

\item{output}{logical, if TRUE, one-step estimator of difference in deviance will be reported.}

\item{...}{other arguments could be passed in.}
}
\value{
list containing of the result of outlier checking.
\item{outliers}{indicator of outlier}
\item{I2}{one-step estimator of difference in deviance}
}
\description{
Identifies outliers influential upon the effects of a covariate 
or a set of covariates for an individual gene.
}

