% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.ZIMD.R
\name{extract.ZIMD}
\alias{extract.ZIMD}
\title{Extracting result from a ZIMD object (two factor test)}
\usage{
extract.ZIMD(obj, get = "contrast", compare = list(A = NULL, B = NULL),
  log2FC.threshold = 0, p.adj = "BY", Log2 = TRUE, ...)
}
\arguments{
\item{obj}{a ZIMD object containing test statistics.}

\item{get}{which one to extract, 'contrast' or a specific covariate.}

\item{compare}{if get='contrast', a comparison of group A versus group B must be specified. 
A and B have to be two of group names in the contrast matrix.}

\item{log2FC.threshold}{log2 fold change threshold for inequality testing. Default is 0. 
That indicates a test that the log2 fold change 
is equal to 0. If log2FC.threshold is specified not equal to 0, 
an inequality test will be performed for |log2FC|>threshold.}

\item{p.adj}{p-value adjustment method, default is 'BY' correction.}

\item{Log2}{logical, if True, difference in mean and dispersion are given in log2 scale.}
}
\value{
a matrix containing testing results.
\item{log2FC.mean}{log2 fold change for mean}
\item{statistic.mean}{test statistic for mean}
\item{pvalue.mean}{p-value for mean}
\item{FDR.mean}{FDR after correction for mean}
\item{log2FC.dispersion}{log2 fold change for dispersion}
\item{statistic.dispersion}{test statistic for dispersion}
\item{pvalue.dispersion}{p-value for dispersion}
\item{FDR.dispersion}{FDR after correction for dispersion}
}
\description{
This program will extract testing results and can simultaneously perform 
a two-sided Wald's test whether log2FC = 0 or an inequality test, in which 
the null hypothesis is that "|log2FC| is less or equal than a given threshold"; 
wehreas the alternative is "|log2FC| is greater than threshold".
}
\seealso{
inequality test is done by inequality.test.
}

