% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.counts.R
\name{filter.counts}
\alias{filter.counts}
\title{Filtering out lowly expressed genes}
\usage{
filter.counts(counts, mean.cpm.cutoff = 0.05, num.per.group = 0,
  threshold.per.group = 1, group = NULL, lib.sizes = NULL,
  remove.zeros = TRUE)
}
\arguments{
\item{counts}{matrix of raw counts.}

\item{mean.cpm.cutoff}{cutoff point for mean counts per million (cpm).}

\item{num.per.group}{If num.per.group > 0, gene will be filtered by
number of samples with cpm > threshold per group. If 
num.per.group = 4, that means gene which's cpm > threshold 
in at least 4 samples will be kept. Note that  
mean.cpm.cutoff is unless when num.per.group > 0. The 
default is num.per.group equals to 0.}

\item{threshold.per.group}{given threshold for cpm per group. Default is 1.}

\item{group}{vector of factors containing group/treatment information}

\item{lib.sizes}{library size for each sample. If not given, it will be 
estimated by the column sum of each sample.}

\item{remove.zeros}{logical, if True (default), 
non-expressed genes will be removed first.}
}
\value{
filtered counts matrix
}
\description{
Filtering out lowly expressed genes
}
\examples{
library(MDSeq)
data(sampleData)

dat <- sample.exprs
dim(dat) 

# group information
group <- sample.pheno$group
 
# lowly expressed gene filtered by mean cpm value across all samples
dat.filtered <- filter.counts(dat, mean.cpm.cutoff = 0.1)  
dim(dat.filtered)

# lowly expressed gene filtered by the number of samples which are 
# equal to or greather than given threshold in each treatment/group
dat.filtered <- filter.counts(dat, num.per.group = 4, threshold.per.group = 1, group = group)  
dim(dat.filtered)

}

