% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.model.matrix.R
\name{get.model.matrix}
\alias{get.model.matrix}
\title{Creating design matrix from contrast matrices and setting contrasts if not supplied}
\usage{
get.model.matrix(group, contrast.matrix = list(mean = NULL, dispersion =
  NULL), contrast.type = "contr.sum")
}
\arguments{
\item{group}{vector of factors containing group/treatment information}

\item{contrast.matrix}{list containing predefined contrast matices for both mean and dispersion}

\item{contrast.type}{if a contrast matix is given, which type of contrast 
will be used in setting contrasts. Default is sum to zero contrasts, 'contr.sum'. 
Also, other contrast types can be used, such as 
"contr.helmert", "contr.poly", "contr.treatment", and "contr.SAS".}
}
\value{
a list containing design matrices for both mean and dispersion
}
\description{
Creating design matrix from contrast matrices and setting contrasts if not supplied
}
\examples{
group <- factor(rep(1:3,4))
get.model.matrix(group)

#treatment contrast - more tranditional glm forming
get.model.matrix(group, contrast.type="contr.treatment")

}

