% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.ZIMD.R
\name{glm.ZIMD}
\alias{glm.ZIMD}
\title{Zero-Inflated Mean-Dispersion Generalized Linear Models}
\usage{
glm.ZIMD(y, X = NULL, U = NULL, contrast = list(mean = NULL, dispersion =
  NULL), initial.pars = list(beta = NULL, gamma = NULL), offsets = NULL,
  constrOptim.maxIter = 100, constrOptim.tol = 1e-05, ZI = TRUE,
  EM.maxIter = 200, EM.tol = 1e-04, control = list())
}
\arguments{
\item{y}{vector contains the counts of one gene}

\item{X}{vector or matrix contains covariates in the mean GLM}

\item{U}{vector or matrix contains covariates in the dispersion GLM, 
which can be different from X.}

\item{contrast}{list containing design matices for mean and dispersion.}

\item{initial.pars}{list giving initial values for parameters of mean and/or dispersion.}

\item{offsets}{numeric vector of same size as y giving offsets for use in log-linear models.}

\item{constrOptim.maxIter}{maximum iterations of the barrier algorithm in constrained optimization.}

\item{constrOptim.tol}{non-negative number; the relative convergence 
tolerance of the barrier algorithm in constrained optimization.}

\item{ZI}{logical, if True (defualt), the effect of zero-inflation will be considered. 
If ZI=False, the effect of zero-inflation will be ignored in computation.}

\item{EM.maxIter}{maximum iteration for EM algorithm.}

\item{EM.tol}{convergence threshold for EM algorithm.}

\item{control}{other control parameters for constrained optimization.}
}
\value{
an object of class 'glmZIMD' containing the following information:
\item{call}{formula}
\item{contrasts}{contrast}
\item{coefficients}{a list containing estimated coefficients from the GLM fits, in the natural log scale}
\item{se}{a list containing standard errors of coefficients}
\item{s}{zero-inflation parameter and standard error}
\item{loglik}{log-likelihood}
\item{hessian}{hessian matrix}
\item{convergence}{an integer code returned by the optimizer. 0 indicates successful completion}
\item{outer.iterations}{number of iterations of the barrier algorithm}
\item{EM.convergence}{EM indicator}
\item{EM.iter}{numbers of EM iterations}
}
\description{
Fit zero-inflated mean-dispersion generalized linear models (GLMs) on 
both the mean and dispersion for an individual gene. Maximum 
likelihood estimations will be calculated via EM algorithm 
incorporated with linearly constrained optimization 
using an adaptive barrier algorithm.
}
\examples{
# simulate 1000 genes, each has 200 observations.
n <- 200; nsim <- 1000; seed <- 1234

# set covariate coefficient for mean and dispersion
beta <- c(5, log(2), 0.5, 0.5); gamma <- c(4, 0, 0.5, 0.5)

# set proportion of excessive-zero
s <- 0.2

# set design matrix for mean and dispersion 
factors <- c('Control', 'Case')
group <- gl(n=2, k=n/2, labels=factors)
design <- get.model.matrix(group, contrast.type="contr.treatment")

# simulate data nsim (genes) X n (observations)
ysim <- sim.ZIMD(n=n, nsim=nsim, beta=beta, gamma=gamma, contrast = design, s=s, seed=seed)

# one example and covariates for mean and dispersion
test <- ysim$y[1,]
X <- ysim$X[,3:4]; colnames(X) = c("X3", "X4")
U <- ysim$U[,3:4]; colnames(U) = c("U3", "U4")

# estimation for a single gene
# consider zero-inflation
fit1 <- glm.ZIMD(test, X=X, U=U, ZI=T)
# ignore zero-inflation
fit2 <- glm.ZIMD(test, X=X, U=U, ZI=F)

fit1
fit2

}

