% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gof.ZI.R
\name{gof.ZI}
\alias{gof.ZI}
\title{Goodness-of-fit test for zero-inflation}
\usage{
gof.ZI(counts, X = NULL, U = NULL, contrast = list(mean = NULL, dispersion
  = NULL), alpha = 0.05, initial.pars = list(beta = NULL, gamma = NULL),
  offsets = NULL, verbose = TRUE, mc.cores = 1, ...)
}
\arguments{
\item{counts}{matrix of count data}

\item{X}{vector or matrix contains covariates in the mean GLM.}

\item{U}{vector or matrix contains covariates in the dispersion GLM.}

\item{contrast}{list containing contrast for mean and dispersion.}

\item{alpha}{threshold of significance level.}

\item{initial.pars}{list giving initial values for parameters of 
mean and/or dispersion for each gene.}

\item{offsets}{numeric vector of same size as rows of counts giving offsets for use in log-linear models.}

\item{verbose}{logical, if True, progress will be reported.}

\item{mc.cores}{integer equals to or greater than 1 (default), indicating how many cores 
will be used in computation. If mc.cores > 0, then parallel processes 
will be used, based on the 'parallel' package.}

\item{...}{other arguments could be passed in.}
}
\value{
a data frame containing testing result of zero-inflation.
\item{ZI}{zero-inflation indicator.
          0, not zero-inflated.
          1, zero-inflated.
          2, error when testing}
\item{pval}{p-value of likelihood ratio test}
}
\description{
This function will perform likelihood ratio test for the presence of technical excess zeros 
at each gene.
}

