% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inequality.test.R
\name{inequality.test}
\alias{inequality.test}
\title{Inequality test (two factors)}
\usage{
inequality.test(est, se, log2FC.threshold = 1)
}
\arguments{
\item{est}{estimated coefficients.}

\item{se}{standard error of estimated coefficients.}

\item{log2FC.threshold}{non-negative value which specifies 
a log2 fold change threshold, default is 1.}
}
\value{
test statistics and p-value for inequality test
}
\description{
This function will perform an inequality test that evaluates log fold-changes beyond a given threshold level. The null hypothesis 
is "|log2FC| is less or equal than given threshold"; and the 
alternative is "|log2FC| is greater than given threshold".
}

