% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.cutoff.R
\name{outlier.cutoff}
\alias{outlier.cutoff}
\title{Finding cutoff point for outlier checking}
\usage{
outlier.cutoff(alpha = 0.05)
}
\arguments{
\item{alpha}{specific significance level. Default is alpha=0.05.}
}
\value{
a cutoff value corresponding to specific significance level.
}
\description{
The cutoff points of outlier checking will be found
 from a variance-gamma distribution.
}
\examples{
# cutoff point at alpha=0.05
cf <- outlier.cutoff(alpha=0.05) 

}

