% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.ZIMD.R
\name{sim.ZIMD}
\alias{sim.ZIMD}
\title{Generating random samples (genes) from ZIMD distribution}
\usage{
sim.ZIMD(n = 200, nsim = 1000, beta, beta.size = 2, beta.prob = c(0.2,
  0.5), gamma, gamma.size = 2, gamma.prob = c(0.2, 0.5), contrast, s,
  tol = 1e-12, max_val = 1e+300, seed = NA)
}
\arguments{
\item{n}{number of total observations in each simulated sample.}

\item{nsim}{number of simulated samples.}

\item{beta}{preset coefficients in the mean GLM}

\item{beta.size}{number of trials (zero or more)}

\item{beta.prob}{probability of success on each trial}

\item{gamma}{preset coefficients in the dispersion GLM}

\item{gamma.size}{number of trials (zero or more)}

\item{gamma.prob}{probability of success on each trial}

\item{contrast}{list containing contrast matrices 
for mean and dispersion, respectively.}

\item{s}{probability of technical excess zeros}

\item{tol}{tolerance for simulation computation.}

\item{max_val}{maximum value for simulated count.}

\item{seed}{random seed}
}
\value{
a list containing random samples and 
         corresponding covariates for mean and dispersion.
\item{y}{matrix containing random samples.}
\item{X}{vector or matrix containing covariates for mean.}
\item{U}{vector or matrix containing covariates for dispersion.}
}
\description{
Generating random samples (genes) from ZIMD distribution
}
\examples{
# simulate 1000 genes, each has 200 observations.
n <- 200; nsim <- 1000; seed <- 123

# set covariate coefficient for mean and dispersion
beta <- c(5, 1, 0.5, 0.5); gamma <- c(4, 0, 0.5, 0.5)

# set proportion of zero-inflation
s <- 0

# set design matrix for mean and dispersion 
factors <- c('Control', 'Case')
group <- gl(n=2, k=n/2, labels=factors)
design <- get.model.matrix(group, contrast.type="contr.treatment")

# simulate data nsim (genes) X n (observations)
ysim <- sim.ZIMD(n=n, nsim=nsim, beta=beta, gamma=gamma, contrast = design, s=s, seed=seed)

}

